requirements.txt
Pillow~=10.4.0
PyGetWindow~=0.0.9
future~=0.18.2

